refactor(select): Extract menu into separate file
